Update to Node v6 compatibility
This commit bumps the compatibility for the plugin to Node 6, the
current Maintenance LTS.  Node v4 is [no longer supported][node], so
it's not that big of a stretch to do this.  That said, it *does*
explicitly break API compatibility, so the Major version had to go up by
one (plus a beta tag so it isn't just installed by default).

[node]: https://github.com/nodejs/Release#readme

Signed-off-by: Jonathan Sifuentes <jayands.dev@gmail.com>